requirements.txt
opencv-python-headless
python-chess
cairosvg
numpy
